requirements.txt
streamlit
pandas
numpy
scikit-learn
tensorflow
yfinance
google-generativeai
pymongo
joblib
python-dotenv
